Improve install script

+++ helpdesk-distribution/src/main/scripts/install.sh
@@ -19,9 +19,14 @@ mkdir -p /opt/symphony/helpdesk-renderer
 mkdir -p /data/symphony/helpdesk-renderer/bin
 mkdir -p /data/symphony/helpdesk-renderer/logs
 
+echo "Stopping services"
+service helpdesk-api stop
+service helpdesk-bot stop
+service helpdesk-renderer stop
+
 echo "Copy application files"
 cp -f helpdesk-api/helpdesk-api.jar /opt/symphony/helpdesk
-cp -f helpdesk-bot/helpdesk.jar /opt/symphony/helpdesk-bot
+cp -f helpdesk-bot/helpdesk-bot.jar /opt/symphony/helpdesk-bot
 cp -f helpdesk-renderer/helpdesk-renderer.jar /opt/symphony/helpdesk-renderer
 
 echo "Copy startup scripts"
@@ -30,11 +35,29 @@ cp -f helpdesk-bot/startup.sh /data/symphony/helpdesk-bot/bin
 cp -f helpdesk-renderer/startup.sh /data/symphony/helpdesk-renderer/bin
 
 echo "Copy YAML descriptor"
-cp -f helpdesk-bot/application.yaml /data/symphony/helpdesk-bot
+yamlfile=/data/symphony/helpdesk-bot/application.yaml
+if [ -f "yamlfile" ]
+then
+    echo "YAML descriptor already exists."
+else
+    echo "Copy YAML descriptor"
+    cp -f helpdesk-bot/application.yaml /data/symphony/helpdesk-bot
+fi
 
 echo "Copy service files"
 cp -f helpdesk-api/helpdesk-api.service /etc/systemd/system
 cp -f helpdesk-bot/helpdesk-bot.service /etc/systemd/system
 cp -f helpdesk-renderer/helpdesk-renderer.service /etc/systemd/system
+
+sleep 5
+
+echo "Starting services"
+systemctl daemon-reload
+
+service helpdesk-api start
+sleep 3
+service helpdesk-renderer start
+sleep 3
+service helpdesk-bot start
 
 echo "Installation complete"
